+++ alc/mixer/mixer_sse.cpp
@@ -230,7 +230,7 @@ void Mix_<SSETag>(const al::span<const float> InSamples, const al::span<FloatBuf
                 step_count = _mm_cvtss_f32(step_count4);
             }
             /* Mix with applying left over gain steps that aren't aligned multiples of 4. */
-            for(uint_fast32_t leftover{min_len&3};leftover;++pos,--leftover)
+            for(size_t leftover{min_len&3};leftover;++pos,--leftover)
             {
                 dst[pos] += InSamples[pos] * (gain + step*step_count);
                 step_count += 1.0f;
@@ -241,7 +241,7 @@ void Mix_<SSETag>(const al::span<const float> InSamples, const al::span<FloatBuf
                 gain += step*step_count;
 
             /* Mix until pos is aligned with 4 or the mix is done. */
-            for(uint_fast32_t leftover{aligned_len&3};leftover;++pos,--leftover)
+            for(size_t leftover{aligned_len&3};leftover;++pos,--leftover)
                 dst[pos] += InSamples[pos] * gain;
         }
         *CurrentGains = gain;
@@ -261,7 +261,7 @@ void Mix_<SSETag>(const al::span<const float> InSamples, const al::span<FloatBuf
                 pos += 4;
             } while(--todo);
         }
-        for(uint_fast32_t leftover{InSamples.size()&3};leftover;++pos,--leftover)
+        for(size_t leftover{InSamples.size()&3};leftover;++pos,--leftover)
             dst[pos] += InSamples[pos] * gain;
     }
 }
